backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-multipart
httpx
python-dotenv
pytest
